requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
django-cors-headers==4.2.0
django-smtp-ssl==1.0
django-ssl-smtp==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
packaging==23.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
